split config logic to seprate file
