+++ packages/bundler/src/runBundler.ts
@@ -1,12 +1,9 @@
-import ow from 'ow'
 import fs from 'fs'
 
 import { Command } from 'commander'
 import { erc4337RuntimeVersion } from '@account-abstraction/utils'
 import { ethers, Wallet } from 'ethers'
-import { BaseProvider } from '@ethersproject/providers'
 
-import { BundlerConfig, bundlerConfigDefault, BundlerConfigShape } from './BundlerConfig'
 import { BundlerServer } from './BundlerServer'
 import { UserOpMethodHandler } from './UserOpMethodHandler'
 import { EntryPoint, EntryPoint__factory } from '@account-abstraction/contracts'
@@ -15,6 +12,7 @@ import { initServer } from './modules/initServer'
 import { DebugMethodHandler } from './DebugMethodHandler'
 import { DeterministicDeployer } from '@account-abstraction/sdk'
 import { isGeth } from './utils'
+import { resolveConfiguration } from './Config'
 
 // this is done so that console.log outputs BigNumber as hex string instead of unreadable object
 export const inspectCustomSymbol = Symbol.for('nodejs.util.inspect.custom')
@@ -27,34 +25,9 @@ const CONFIG_FILE_NAME = 'workdir/bundler.config.json'
 
 export let showStackTraces = false
 
-export function resolveConfiguration (programOpts: any): BundlerConfig {
-  let fileConfig: Partial<BundlerConfig> = {}
-
-  const commandLineParams = getCommandLineParams(programOpts)
-  const configFileName = programOpts.config
-  if (fs.existsSync(configFileName)) {
-    fileConfig = JSON.parse(fs.readFileSync(configFileName, 'ascii'))
-  }
-  const mergedConfig = Object.assign({}, bundlerConfigDefault, fileConfig, commandLineParams)
-  console.log('Merged configuration:', JSON.stringify(mergedConfig))
-  ow(mergedConfig, ow.object.exactShape(BundlerConfigShape))
-  return mergedConfig
-}
-
-function getCommandLineParams (programOpts: any): Partial<BundlerConfig> {
-  const params: any = {}
-  for (const bundlerConfigShapeKey in BundlerConfigShape) {
-    const optionValue = programOpts[bundlerConfigShapeKey]
-    if (optionValue != null) {
-      params[bundlerConfigShapeKey] = optionValue
-    }
-  }
-  return params as BundlerConfig
-}
-
-export async function connectContracts (
+export async function connectContracts(
   wallet: Wallet,
-  entryPointAddress: string): Promise<{ entryPoint: EntryPoint}> {
+  entryPointAddress: string): Promise<{ entryPoint: EntryPoint }> {
   const entryPoint = EntryPoint__factory.connect(entryPointAddress, wallet)
   return {
     entryPoint
@@ -67,13 +40,13 @@ export async function connectContracts (
  * @param argv
  * @param overrideExit
  */
-export async function runBundler (argv: string[], overrideExit = true): Promise<BundlerServer> {
+export async function runBundler(argv: string[], overrideExit = true): Promise<BundlerServer> {
   const program = new Command()
 
   if (overrideExit) {
     (program as any)._exit = (exitCode: any, code: any, message: any) => {
       class CommandError extends Error {
-        constructor (message: string, readonly code: any, readonly exitCode: any) {
+        constructor(message: string, readonly code: any, readonly exitCode: any) {
           super(message)
         }
       }
@@ -101,7 +74,7 @@ export async function runBundler (argv: string[], overrideExit = true): Promise<
 
   console.log('command-line arguments: ', program.opts())
 
-  const config = resolveConfiguration(programOpts)
+  const [config, provider, wallet] = await resolveConfiguration(programOpts)
   if (programOpts.createMnemonic != null) {
     const mnemonicFile = config.mnemonic
     console.log('Creating mnemonic in file', mnemonicFile)
@@ -112,18 +85,6 @@ export async function runBundler (argv: string[], overrideExit = true): Promise<
     fs.writeFileSync(mnemonicFile, newMnemonic)
     console.log('created mnemonic file', mnemonicFile)
     process.exit(1)
-  }
-  const provider: BaseProvider =
-    // eslint-disable-next-line
-    config.network === 'hardhat' ? require('hardhat').ethers.provider :
-      ethers.getDefaultProvider(config.network)
-  let mnemonic: string
-  let wallet: Wallet
-  try {
-    mnemonic = fs.readFileSync(config.mnemonic, 'ascii').trim()
-    wallet = Wallet.fromMnemonic(mnemonic).connect(provider)
-  } catch (e: any) {
-    throw new Error(`Unable to read --mnemonic ${config.mnemonic}: ${e.message as string}`)
   }
 
   const {
